TASK #73 - Turn JWT endpoint auth to middleware
Created middleware to wrap around endpoints to allow us to verify tokens on protected endpoints.

See /example/hello_world endpoint for an example.

When JWTMiddleware() is wrapped around an endpoint, it will verify the token and parse the userID and username. This is then passed into the next handler using Context. If the token is invalid, it will send back an 401 unauthorized.

To get the JWT token, you should register a user or login